// src/app/models/DetailMonthlyPlanningCuring.ts
export class DetailMonthlyPlanningCuring {
  detailIdCuring: string;
  docNumber: string;
  partNumber: number;
  total: number;
  netFulfilment: number;
  grossReg: number;
  netReq: number;
  reqAhmOem: number;
  reqAhmRem: number;
  reqFdr: number;
  differenceOfs: number;
  status: number;
  }
  