// src/app/models/DetailShiftMonthlyPlanCuring.ts
export class DetailShiftMonthlyPlanCuring {
    detailShiftIdCuring: number;
    detailDailyIdCuring: number;
    moId: string;
    mpCuringId: string;
    actualDate: Date;
    itemCuring: string;
    partNumber: number;
    workCenterText: string;
    cavity: number;
    cavityUsage: number;
    whShift1: number;
    whShift2: number;
    whShift3: number;
    kapaShift1: number;
    kapaShift2: number;
    kapaShift3: number;
    totalKapa: number;
    cavityExist: number;
    status: number;

    shift: number;
    changeDate: Date;
    wct: string;
    }
    